+++ apps/site-nft/internals/webpack.ts
@@ -1,7 +1,5 @@
 import { getConfig } from "@thecointech/site-base/internal/webpack";
 
-const config = await getConfig([], {
+export default getConfig([], {
   port: 3003
 });
-
-export default config;