+++ libs/site-prismic/src/slices/index.ts
@@ -7,6 +7,7 @@ export const components = {
   captioned_image: dynamic(() => import("./CaptionedImage")),
   ccqi_category: dynamic(() => import("./CcqiCategory")),
   ccqi_score: dynamic(() => import("./CcqiScore")),
+  comparison_block: dynamic(() => import("./ComparisonBlock")),
   hero: dynamic(() => import("./Hero")),
   project_details: dynamic(() => import("./ProjectDetails")),
   rich_text: dynamic(() => import("./RichText")),