+++ apps/harvester/CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.6](https://github.com/thecointech/thecoin/compare/v0.6.6-beta.0...v0.6.6) (2026-08-06)
+
+**Note:** Version bump only for package harvester
+
+
+
+
+
+## [0.6.6-test.7](https://github.com/thecointech/thecoin/compare/v0.6.6-test.6...v0.6.6-test.7) (2026-08-05)
+
+**Note:** Version bump only for package harvester
+
+
+
+
+
+## [0.6.6-test.1](https://github.com/thecointech/thecoin/compare/v0.6.6-test.0...v0.6.6-test.1) (2026-07-30)
+
+**Note:** Version bump only for package harvester
+
+
+
+
+
+## [0.6.6-test.0](https://github.com/thecointech/thecoin/compare/v0.6.5...v0.6.6-test.0) (2026-07-30)
+
+**Note:** Version bump only for package harvester
+
+
+
+
+
+
+
+**Note:** Version bump only for package harvester
+
+
+
+
+
+## [0.6.5-test.0](https://github.com/thecointech/thecoin/compare/v0.6.4...v0.6.5-test.0) (2026-07-23)
+
+
+### Bug Fixes
+
+* defensive fix to ensure we don't run into the same issue later. ([a708110](https://github.com/thecointech/thecoin/commit/a708110f4eceafdaf44ce84a9bdcd85c8ce3a9ea))
+
+
+
+
+
+## [0.6.4](https://github.com/thecointech/thecoin/compare/v0.6.4-test.1...v0.6.4) (2026-07-22)
+
+**Note:** Version bump only for package harvester
+
+
+
+
+
+## [0.6.4-test.1](https://github.com/thecointech/thecoin/compare/v0.6.4-test.0...v0.6.4-test.1) (2026-07-01)
+
+**Note:** Version bump only for package harvester
+
+
+
+
+
+## [0.6.4-test.0](https://github.com/thecointech/thecoin/compare/v0.6.3...v0.6.4-test.0) (2026-07-01)
+
+**Note:** Version bump only for package harvester
+
+
+
+
+
 ## [0.6.3](https://github.com/thecointech/thecoin/compare/v0.6.3-test.0...v0.6.3) (2026-05-05)
 
 **Note:** Version bump only for package harvester